Delay the canonicalization until after processing the CLI arguments
